songbird/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
beautifulsoup4==4.12.2
black==23.9.1
Brotli==1.1.0
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
coverage==5.5
cssselect==1.2.0
deprecation==2.1.0
eyed3==0.9.7
fake-useragent==1.2.1
fastapi==0.103.2
ffprobe-python==1.0.3
filetype==1.2.0
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
isort==5.12.0
lxml==4.9.3
mutagen==1.47.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
parse==1.19.1
pathspec==0.11.2
platformdirs==3.10.0
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.19.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pydub==0.25.1
pyee==8.2.2
pyparsing==3.1.1
pyppeteer==1.0.2
pyquery==2.0.0
python-dotenv==1.0.0
python-magic==0.4.27
requests==2.31.0
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
toml==0.10.2
tqdm==4.66.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
w3lib==2.1.2
websockets==10.4
yt-dlp==2023.7.6
zipp==3.16.2
